Replace two bools with uint8, static set/get funcs
The setter/getter functions add a bit of "boiler plate" and could be
smaller; clarity was chosen over conciseness, which stands out in
the usage code.

As expected ( and checked with https://godbolt.org/ ) the static
functions vanish and the single flag struct member is smaller than two
bool members.